requirements.txt
# Specify Python version for PyNV compatibility
# Python version: 3.10.7 "pyenv local 3.10.7"

aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.12.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.1
chromadb==0.4.0
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.99.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
greenlet==3.2.1
h11==0.16.0
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.306
langsmith==0.0.92
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.4
onnxruntime==1.19.2
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
posthog==4.0.1
propcache==0.3.1
protobuf==6.30.2
pulsar-client==3.5.0
pydantic==1.10.22
pydantic_core==2.33.2
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.27.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
